Add user account editing
Add a form to edit users in the backend. Add the necessary language lines and institute a basic email system with email layouts for HTML and plain-text emails. Add an email that can be sent to users containing their new password. Also add the main script file to the internal and backend layouts.