Lock out inactive users
Prevent inactive users from accessing anything in the internal area (they can still access the backend if they have that privilege so as not to accidentally lock out the superuser). Prepare for forced password changes by adding some translations and a TODO in the Authenticate middleware. First we need to make the password change page.