Add a Member model, pull in Revisionable
Add a Member model to hold all information about the fraternity member. This model is linked to the user account. There are a few changes to the layout and the authentication process necessary to reflect this new structure.

Also pull in VentureCraft/revisionable to manage revision history on the members (and other models) in the future.