Describe settings in JSON file
Describe all the available system settings in resources/settings.json. Use this file to feed the Backend -> System Settings controller and view as well as the GlobalSettingsSeeder. The settings.json is ready to contain other settings groups besides "system", most notably default settings for individual users. These will be added as needed. Adding a new system setting is now as easy as adding a new element to the JSON file without needing to touch the controller or view. For now, the supported types of setting are string, csv, and country; more types may be added as needed.