Add addresses and phone numbers to history
When a new address or phone number is added to a member, or when one of them is deleted, add a new revision to the member. This is better than having a separate revision history for addresses and phone numbers because no additional steps need to be taken for the notification emails to be sent when this vital information changes.

Also make minor improvements to the version history display and add some missing translations that were hardcoded until now.