Redesign user/member management
Fix some serious issues with user and member management and the relationships between users and members. With this commit, deleting and purging members and users works flawlessly even if the two models are connected. Some functionality is moved into a new MemberRepository, e.g. creating an array ready to be fed into a bsSelect to select a member from a list. Also add some missing translations in this area.